requirements.txt
notion-py==0.0.10
discord==1.7.3
discord.py==1.7.3